chore: split out tauri commands for readability (#1150)
moved all our tauri commands to their own file 

1. for readability and
2. because i did that in the V2 branch and i think i've missed changes
when merging when there are conflicts 😬
